Fix for the FishEyeNoInverse regression in CSSD100014407 (r44245 changed the heuristic dominant type tie-breaking which exposed a different bug)
CR: Ran

Pre-commit: http://cvcc-fc14-01.iil.intel.com:8080/job/PersonalCompilerReport/1457/

+++ opencl-intel/backend/vectorizer/Heuristic/InstCounter.cpp
@@ -222,7 +222,7 @@ Type* WeightedInstCounter::estimateDominantType(Function &F, DenseMap<Loop*, int
   // which, in turn, depends on the values of the Type pointers, which change
   // from run to run. This makes the choice inconsistent between runs.
   Type* DominantType = Type::getInt32Ty(F.getContext());
-  int MaxCount = 0;
+  float MaxCount = 0;
   for (DenseMap<Type*, float>::iterator I = countMap.begin(), 
        E = countMap.end(); I != E; ++I) {
     if ((I->second > MaxCount) ||
